# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Customer booking service API
  version: 1.0.0
  description: API for managing customer bookings

tags:
  - name: Customer
    description: Operations related to customers
  - name: Brand
    description: Operations related to brands
  - name: Booking
    description: Operations related to bookings

paths:
  #Customer
  /api/customer/all:
    get:
      summary: Get all customers
      tags:
        - Customer
      responses:
        '200':
          description: A list of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerDto'

  /api/customer/get:
    get:
      summary: Get a customer by ID
      tags:
        - Customer
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDto'
        '404':
          description: Customer not found

  /api/customer/create:
    post:
      summary: Create a new customer
      tags:
        - Customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerDto'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDto'

  /api/customer/update:
    put:
      summary: Update a customer
      tags:
        - Customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerDto'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDto'
        '404':
          description: Customer not found

  /api/customer/delete:
    delete:
      summary: Delete a customer by ID
      tags:
        - Customer
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Customer deleted
        '404':
          description: Customer not found

  /api/customer/bookings:
    get:
      summary: Get bookings for customer
      tags:
        - Customer
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: A list of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookingDto'

  #Brand
  /api/brand/all:
    get:
      summary: Get all brands
      tags:
        - Brand
      responses:
        '200':
          description: A list of brands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BrandDto'

  /api/brand/get:
    get:
      summary: Get a brand by ID
      tags:
        - Brand
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Brand found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandDto'
        '404':
          description: Brand not found

  /api/brand/create:
    post:
      summary: Create a new brand
      tags:
        - Brand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBrandDto'
      responses:
        '201':
          description: Brand created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandDto'

  /api/brand/update:
    put:
      summary: Update a brand
      tags:
        - Brand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBrandDto'
      responses:
        '200':
          description: Brand updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandDto'
        '404':
          description: Brand not found

  /api/brand/delete:
    delete:
      summary: Delete a brand by ID
      tags:
        - Brand
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Brand deleted
        '404':
          description: Brand not found

  /api/brand/bookings:
    get:
      summary: Get bookings for brand
      tags:
        - Brand
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: A list of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookingDto'

  #Booking
  /api/booking/create:
    post:
      summary: Creates booking for customer
      tags:
        - Booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingDto'
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDto'
        '404':
          description: Customer not found

  /api/booking/delete:
    delete:
      summary: Deletes booking for customer
      tags:
        - Booking
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Booking deleted
        '404':
          description: Booking not found

  /api/booking/addBrand:
    patch:
      summary: Adds brand to booking
      tags:
        - Booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddBrandDto'
      responses:
        '204':
          description: Brand added
        '404':
          description: Booking or Brand not found

components:
  schemas:
    CustomerDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        fullName:
          type: string
          example: John
        email:
          type: string
          example: John@gmail.com
        status:
          type: string
          example: ACTIVE
        age:
          type: integer
          example: 27
        created:
          type: string
          format: date
          example: "2024-01-15"
        updated:
          type: string
          format: date
          example: "2024-01-15"
    CreateCustomerDto:
      type: object
      properties:
        fullName:
          type: string
          example: John
        email:
          type: string
          example: John@gmail.com
        status:
          type: string
          example: ACTIVE
        age:
          type: integer
          example: 27
    UpdateCustomerDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        fullName:
          type: string
          example: John
        email:
          type: string
          example: John@gmail.com
        status:
          type: string
          example: ACTIVE
        age:
          type: integer
          example: 27
    BrandDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: John
        address:
          type: string
          example: Street example
        created:
          type: string
          format: date
          example: "2024-01-15"
        updated:
          type: string
          format: date
          example: "2024-01-15"
        shortCode:
          type: string
          example: BYD
    CreateBrandDto:
      type: object
      properties:
        name:
          type: string
          example: John
        address:
          type: string
          example: Street example
        created:
          type: string
          format: date
          example: "2024-01-15"
        updated:
          type: string
          format: date
          example: "2024-01-15"
        shortCode:
          type: string
          example: BYD
    UpdateBrandDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: John
        address:
          type: string
          example: Street example
        created:
          type: string
          format: date
          example: "2024-01-15"
        updated:
          type: string
          format: date
          example: "2024-01-15"
        shortCode:
          type: string
          example: BYD
    BookingDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: Booking title
        description:
          type: string
          example: Booking description
        status:
          type: string
          example: CANCELLED
        recordCreated:
          type: string
          format: date
          example: "2024-01-15"
        recordUpdated:
          type: string
          format: date
          example: "2024-02-15"
        startDate:
          type: string
          format: date
          example: "2024-03-15"
        endDate:
          type: string
          format: date
          example: "2024-04-15"
        brandName:
          type: string
          example: BYD
        customerId:
          type: integer
          format: int64
          example: 1
    CreateBookingDto:
      type: object
      properties:
        title:
          type: string
          example: Booking title
        description:
          type: string
          example: Booking description
        status:
          type: string
          example: CANCELLED
        startDate:
          type: string
          format: date
          example: "2024-03-15"
        endDate:
          type: string
          format: date
          example: "2024-04-15"
        customerId:
          type: integer
          format: int64
          example: 1
    AddBrandDto:
      type: object
      properties:
        brandId:
          type: integer
          format: int64
          example: 1
        customerId:
          type: integer
          format: int64
          example: 2